Deleted more toxic code (things are getting out of hand).
With more and more time it seems as if the issue of toxic code is
getting worse. I just had to deal with TRIPLICATES of code and more. I
might have deleted several things that I was not suppose to delete. If I
did, I apologize, but it just keeps getting worse and I don't know how
to manage it anymore really.